fixed priority queue, fixed getMapBounds
