+++ package.json
@@ -13,17 +13,17 @@
     },
     "license": "UNLICENSED",
     "devDependencies": {
-        "@timohausmann/quadtree-ts": "^2.0.0-beta.1",
         "@types/node": "^14.17.32",
         "prettier": "3.0.3",
         "typescript": "^4.3.5"
     },
     "dependencies": {
         "@chronodivide/game-api": "^0.45.0",
+        "@datastructures-js/priority-queue": "^6.3.0",
+        "@timohausmann/quadtree-ts": "^2.0.0-beta.1",
         "@types/lodash": "^4.14.199",
         "@types/luxon": "^3.3.2",
         "lodash": "^4.17.21",
-        "luxon": "^3.4.3",
-        "priority-queue-typescript": "^1.0.1"
+        "luxon": "^3.4.3"
     }
 }
