+++ src/bot/logic/map/map.ts
@@ -1,33 +1,19 @@
-import { GameApi, MapApi, PlayerData, Point2D, UnitData } from "@chronodivide/game-api";
+import { GameApi, MapApi, PlayerData, Point2D, Tile, UnitData } from "@chronodivide/game-api";
+import _ from "lodash";
+
+const MAX_WIDTH_AND_HEIGHT = 500;
 
 // Expensive one-time call to determine the size of the map.
 // The result is a point just outside the bounds of the map.
 export function determineMapBounds(mapApi: MapApi): Point2D {
-    // TODO Binary Search this.
-    // Start from the last spawn positions to save time.
-    let maxX: number = 0;
-    let maxY: number = 0;
-    mapApi.getStartingLocations().forEach((point) => {
-        if (point.x > maxX) {
-            maxX = point.x;
-        }
-        if (point.y > maxY) {
-            maxY = point.y;
-        }
-    });
-    // Expand outwards until we find the bounds.
-    for (let testX = maxX; testX < 10000; ++testX) {
-        if (mapApi.getTile(testX, 0) == undefined) {
-            maxX = testX;
-            break;
-        }
-    }
-    for (let testY = maxY; testY < 10000; ++testY) {
-        if (mapApi.getTile(testY, 0) == undefined) {
-            maxY = testY;
-            break;
-        }
-    }
+    // Probably want to ask for an API change to get this.
+    // Note that the maps is not always a rectangle!
+    const zeroTile = { rx: 0, ry: 0 } as Tile;
+    const allTiles = mapApi.getTilesInRect(zeroTile, { width: MAX_WIDTH_AND_HEIGHT, height: MAX_WIDTH_AND_HEIGHT });
+
+    const maxX = _.maxBy(allTiles, (tile) => tile.rx)?.rx!;
+    const maxY = _.maxBy(allTiles, (tile) => tile.ry)?.ry!;
+
     return { x: maxX, y: maxY };
 }
 
@@ -35,7 +21,7 @@ export function calculateAreaVisibility(
     mapApi: MapApi,
     playerData: PlayerData,
     startPoint: Point2D,
-    endPoint: Point2D
+    endPoint: Point2D,
 ): { visibleTiles: number; validTiles: number } {
     let validTiles: number = 0,
         visibleTiles: number = 0;
@@ -60,7 +46,7 @@ export function getPointTowardsOtherPoint(
     endLocation: Point2D,
     minRadius: number,
     maxRadius: number,
-    randomAngle: number
+    randomAngle: number,
 ): Point2D {
     let radius = minRadius + Math.round(gameApi.generateRandom() * (maxRadius - minRadius));
     let directionToSpawn = Math.atan2(endLocation.y - startLocation.y, endLocation.x - startLocation.x);
